AndroidDriver is now built into the standalone server.
A request using the normal Selenium bindings for DesiredCapabilities.Android will install the 'android-driver-app' which is a simple app that
contains only a WebView and nothing else. The server will start the app and auto switch to the WEBVIEW.

A new option has been added -noWebviewApp this should be added to the command line to prevent the extraction and resigning of the android-driver-app,
will speed up start time of the standalone server if you don't require the android web driver.

Default standalone server no longer requires at least one app specified since the android-driver-app is always bundled with it.