Get a more detailed error message when instrumentation fails.
The way this works is that when we get an error from
'adb shell am instrument', we try to rerun the same command
with '-w' (wait for instrumentation to finish) so we get
more details.

We don't want to always start instrumentation with -w because
this will timeout in the success case.

We append the '-w' right after 'shell am instrument'. Appending
it to the end of command doesn't appear to be working.