Remove SelendroidKeys.ANDROID_HOME as Android doesn't support this
Summary:
Remove `SelendroidKeys.HOME`. The fact that it is available is confusing because `Instrumentation.sendKeyDownUpSync(KeyEvent.KEYCODE_HOME)` is a no-op on Android for security reasons.

Any code that depends on this is not working and therefore it should be OK to remove it even though it is a breaking API change.

To make this cleare, throw an exception in case someone manages to send `KeyEvent.KEYCODE_HOME` manually.

Test Plan:
- `mvn clean install`
- Tried `driver.getKeyboard().sendKeys(String.valueOf('\uE101'))` and the correct explanatoy error response was returned.

Reviewers: jimpurbrick

Reviewed By: jimpurbrick

Differential Revision: https://reviews.facebook.net/D24249