Make retries in standalone driver configurable
Summary:
The value for the amount of retries for starting an emulator was previously
hardcoded. This diff adds an aption to the SelendroidConfiguration to configure
the amount of retries and adds test coverage for the retry feature.

For reference, this was the commit that introduced the retry logic:
https://github.com/selendroid/selendroid/commit/a12e5038f80fbed0a208d02a5f4e5926ee627e61

Test Plan: - run the tests (especially the SelendroidStandaloneDriverTest)

Reviewers: mkonicek

Reviewed By: mkonicek

Subscribers: dstroganova

Differential Revision: https://reviews.facebook.net/D28341