Elements don't need to be enabled to be visible.
The WebDriver protocol has a separate facility to
check whether elements are enabled, called `isEnabled`:
https://dvcs.w3.org/hg/webdriver/raw-file/default/webdriver-spec.html

I ran into this issue because we had a button that was disabled
and would only get enabled as we type a username into an `EditText`.
The test would fail because it did some basic checks to see whether
the UI looked good, including checking that the button was visible.