Adding "model" as a configurable parameter option.
So users can select a specific device model together with the SDK version
for their tests. Previously, users can only specify the SDK version
for the tests and if there are several devices with the same SDK
version, selendroid will automatically select the first device to
carry out the test, which may not be the intended device for the
test.

This parameter can be ignored if the users are not desired to select
a specific device model for the test.

Signed-off-by: Binh Vu <bvu@paypal.com>