[client] Fix SelendroidDriver#readCallLog()
AddCallLogTest was broken. Turns out the `readCallLog()` method was
actually broken. The `execute()` method returns an `ArrayList<String>`,
so the cast to `String` was erroring.

Changed the cast to the appriopriate one and fixed the loop that parses
the individual call log entries. Added extra error handling for the
class cast exception.