[test-app] Fix BackgroundAppTest
The test looks for the lines containing "mCurrentFocus" or "mFocusedApp"
in the output of an adb command to figure out which app is currently in
the foreground. Piping the ouput of the command through grep isn't
reliable. On my laptopt it yields an sh error. Moved the line filtering
to plain java instead.