[server] Do not search in multiple top level views
Summary: In some cases selendroid ended up looking for elements in
multiple top level views. This resulted in cases where we would look up
an element by id, but the element was present in in a top level view
that was not a child of the current root view. The server happily
returns the element, but as soon as you try to interact with the
element, the known elements cache will check if the element is part of
the root level view and if that's not the case throw a
`StaleElementReference` without further information.

Instead of this confusing behavior, selendroid should search only in the
current top level view.

Test Plan:
- Ran our internal e2e tests against selendroid with this patch.
- unit tests