need to "Proxy" WebChromeClient / WebViewClient
When switching to a webview selendroid will call webView.set*Client.
This overrides any existing Client that may have been set in the AUT.
This is undesired behavior as the webView will no longer act the same
as it would when not under control of selendroid.

Adding a 'Wrapper' for the existing Web*Client's that may have been set.

Could not use a Java Proxy, as the Web*Client classes are concrete and there
is no interface to Proxy. Other options to supposedly proxy a concrete class
involve byte code manipulation, which aren't the same on Android compared to
a standard jvm. I looked at a couple of library's to attempt to use them,
but I could not make them work satisfactorily.

Adding a sample test that shows the test-app having it's own Web*Client that in turn
causes the app to behave 'differently' than when selendroid overrode the Web*Client.