wat the wrap.
Need to wrap the cordova clients as well.

My test app also likes to do "special" things to the cordova webview class.
The WebView's class name is that of a nested class and doesn't return anything for getSimpleName,
so need to rely on getting a ClassCast exception to determine if it really is cordova or not.