Added support to keypass switch for signing the AUT wit custom keystore with keypass enabled.
Changes:
added a switch -keyPassword, set and get methods in SelendroidConfiguration.
Modified androidDebugKeystore to get read the keypass from SelendroidConfiguration. 
added command line arugments in signTestServer method to include keypass if required.
added  two tests to SelendroidServerBuilderTest, for signing test app
and server app to sign with custom keystore with keypass enabled.

Tests run:
Units were successful.
Other tests with test apps:
Tested with default keystore:
from command line with app installed and not installed,
with in the script with app installed and not installed.
Tested with custom keystore:
from command line with app installed and not installed
with in the script with app installed and not installed.
Tested with custom keystore with keypass:
from command line with app installed and not installed
with in the script with app installed and not installed.