Add support for launching implicit intents
Summary:
Sometimes a test may want to launch an Activity via URI (e.g. to encode
a user ID or other information). This adds an "intent_uri" option to
both lightweight and regular server instrumentation that acts as a
fallback in case the main activity name is not specified.

Test Plan: Attempting to compile locally produces a stack overflow in javac. Not sure what else I can do here.

Reviewers: asm

Subscribers: dancol

Differential Revision: https://reviews.facebook.net/D56949